Github #5: cleanup develop-eggs needs same params as bootstrap
In the previous commit, I forgot to pass '**step_kw'. As a result, in release
builds (packaging subfactory), the first call to cleanup occurs in the
'build' subdirectory instead of 'src'.
This has a side-effect: creating 'build' in the first place.
Then, the 'mv' step that renames the extracted directory as 'build' for
final testing actually moves it below 'build', and everything fails