Stop using shallow-clones by default + doc effort on options
We have found --depth=1 to be harmful, refetching everything, and having
our Git server block the builds.
Actually, the disk space gain of shallow clones for Odoo is low enough
(about a division by two) that it isn't worth the effort to investigate
so much.

Still keeping it as an option with --depth=2, hoping it'd be better.