Applying build-requires at build time
If a slave has two versions of a given capability, the step that fixes
execution properties (bin/lib directories, ports, etc.) cannot choose which
set of options to choose.
This was a bit silly if that slave got selected
through a build-requires statement that has the effect to
restrict applicable capabilities on that slave to precisely one.

We did not stumble on this until right now, because we've had
either capabilities that were used in build-requires
or in build-for (juste one, actually for build-for), but now we have some
(typically the python capability).